Add staging runtime API
Quoting @rphmeier (with small edits):

The broader idea I have around runtime API upgrades is that we'd
maintain a vstaging which is based on the latest API version and has a
special STAGING_VERSION (u32::MAX). We'd be able to conditionally enable
this on runtimes for testing. And then we'd eventually set the next
runtime API version to be equal to vstaging and start a fresh vstaging.